+++ projects/workspace/src/app/custom-step/custom-step.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { NgFlowchartStepComponent } from 'projects/ng-flowchart/src/lib/ng-flowchart-step/ng-flowchart-step.component';
-import { NgFlowchart } from 'projects/ng-flowchart/src/public-api';
+import { NgFlowchart } from 'projects/ng-flowchart/src';
 import { RouteStepComponent } from './route-step/route-step.component';
 
 @Component({
